Feature: self-upgrade: rename module self-update -> self-upgrade
