+++ src/self_upgrade/checker.rs
