+++ src/self_upgrade/downloader.rs
