+++ src/self_upgrade/environment.rs
