+++ src/self_upgrade/mod.rs
