+++ src/self_upgrade/replacer.rs
